added routing for my recipes
