+++ server/routes/recipe_router.py
@@ -3,6 +3,8 @@ from dotenv import load_dotenv
 import os
 import httpx
 import asyncio
+import aiohttp
+from aiohttp_client_cache import CachedSession, SQLiteBackend
 
 router = APIRouter(
     prefix='/recipes',
@@ -10,6 +12,11 @@ router = APIRouter(
     responses={404: {"description": "Not Found"}}
 )
 
+cache = SQLiteBackend(
+    cache_name='local_cache',
+    expire_after=3600,
+    ignored_params=['apiKey'] # uses cache response even if the token differs
+)
 
 load_dotenv(os.path.dirname(os.getcwd())+'/.env')
 
@@ -32,3 +39,26 @@ async def basic_get():
     # data = asyncio.run(r)
     # return info
     return {'data': r.json()} # json to serialize it upon return
+
+async def get_api(session, url, params):
+    #async with session.get(url + f'?apiKey={SPOONACULAR_API_KEY}&query=tacos') as response:
+    #    return await response.text()
+    async with session.get(url, params=params) as response:
+        return await response.json() # returns dictionary of the results.
+
+@router.get('/aihttp-version')
+async def basic_get():
+    # make api call - couldn't we just open the session in a dependency? like with get_db?
+    # that way we don't have to write this code more than once
+    async with CachedSession(cache=cache) as session:
+        #async with aiohttp.ClientSession() as session:
+        response = await get_api(
+            session, 
+            'https://api.spoonacular.com/recipes/complexSearch', 
+            {"apiKey": SPOONACULAR_API_KEY, "query": "tacos"})
+        # print(response.get('results')[0])
+    return {'data': response} # json to serialize it upon return
+
+@router.post('/')
+async def create_cart():
+    pass
